Split enums to seperate classes
Remove need for class.
Update nfproj for individual classes.
